+++ source/extensions/isaacsim.streaming.rtsp/docs/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## [0.1.3] - 2026-04-27
+
+### Added
+- Pass configured SRTX sensor-set names from `RTSPCameraHelper` to `RTSPStreamWriter` and into the `LdrColor` annotator `init_params`. Resolution is delegated to `omni.replicator.srtx.resolve_sensor_set_name_for_render_product()`.
+- Optional dependency on `omni.replicator.srtx` for sensor-set resolution.
+
 ## [0.1.2] - 2026-04-24
 
 ### Fixed